skeleton for analytics selector
