+++ front-end/package.json
@@ -31,6 +31,7 @@
     "@radix-ui/react-popover": "^1.1.6",
     "@radix-ui/react-portal": "^1.1.4",
     "@radix-ui/react-presence": "^1.1.2",
+    "@radix-ui/react-radio-group": "^1.3.4",
     "@radix-ui/react-roving-focus": "^1.1.2",
     "@radix-ui/react-separator": "^1.1.2",
     "@radix-ui/react-slot": "^1.1.2",
