+++ front-end/src/app/(private)/home/[userId]/settings/page.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function page() {
+  return (
+    <div className='flex items-center justify-center'>
+      <div>Settings</div>
+    </div>
+  )
+}
