--- front-end/src/app/(settings)/settings/page.tsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export default function page() {
-  return (
-    <div className='flex items-center justify-center'>
-      <div>Settings</div>
-    </div>
-  )
-}
